Sorted classes, exceptions and enumerations into packages
