+++ src/main/java/classes/Event.java
@@ -1,33 +1,41 @@
-public class ToDo extends Task {
-    public ToDo(String description) {
+package classes;
+
+public class Event extends Task {
+    private String dateAndTime;
+
+    public Event(String description, String dateAndTime) {
         super(description);
-        setSymbol();
+        this.dateAndTime = dateAndTime;
     }
 
     public void setDone() {
         this.done = true;
         System.out.println("Nice! I've marked this task as done:");
-        System.out.println("[T][✓] " + this.description);
+        System.out.println("[D][✓] " + this.description + " (at: " + getDateAndTime() + ")");
     }
 
     public void printAdd(int storeSize) {
         System.out.println("Got it. I've added this task:");
-        System.out.println(getSymbol() + "[✗] " + getDescription());
+        System.out.println(getSymbol() + "[✗] " + getDescription() + " (at: " + getDateAndTime() +")");
         System.out.println("Now you have " + (storeSize) + " tasks in the list");
     }
 
     public void printDelete(int storeSize) {
         System.out.println("Noted. I've removed this task:");
-        System.out.println(getSymbol() + getDone() + " " + getDescription());
+        System.out.println(getSymbol() + getDone() + " " + getDescription() + " (at: " + getDateAndTime() +")");
         System.out.println("Now you have " + (storeSize) + " tasks in the list");
     }
 
     public void setSymbol() {
-        symbol = "[T]";
+        symbol = "[E]";
     }
 
     public String getSymbol() {
         return symbol;
     }
 
+    @Override
+    public String getDateAndTime() {
+        return dateAndTime;
+    }
 }
