+++ src/main/java/classes/Task.java
@@ -1,3 +1,5 @@
+package classes;
+
 public abstract class Task {
     protected String description;
     protected boolean done;
