+++ src/main/java/enumerations/CommandEnum.java
@@ -1,3 +1,5 @@
+package enumerations;
+
 public enum CommandEnum {
     TODO, DEADLINE, EVENT, LIST, DELETE, DONE, BYE
 }
