+++ src/main/java/exceptions/Exception.java
@@ -1,2 +1,4 @@
+package exceptions;
+
 public class Exception extends Throwable {
 }
