+++ src/main/java/exceptions/NoDeadlineException.java
@@ -1,2 +1,4 @@
+package exceptions;
+
 public class NoDeadlineException extends Exception {
 }
