+++ src/main/java/exceptions/NoDescriptionException.java
@@ -1,2 +1,4 @@
+package exceptions;
+
 public class NoDescriptionException extends Exception {
 }
