+++ src/main/java/exceptions/NoTimeframeException.java
@@ -1,2 +1,4 @@
+package exceptions;
+
 public class NoTimeframeException extends Exception {
 }
